\name{ellipticity}
\alias{ellipticity}
\alias{ellipseEccentricity}
\alias{ellipseFocus}
\alias{ellipseRa}
\alias{ellipseRp}
\alias{ellipse.l}
\title{Formulas for the ellipse}
\description{\code{ellipticity} ellipticity = flattening factor
\code{ellipseEccentricity} eccentricity of the ellipse
\code{ellipseFocus} focus of the ellipse
\code{ellipseRa} radius at apoapsis (the farthest distance)
\code{ellipseRp} radius at periapsis (the closest distance)
\code{ellipse.l} semi-latus rectum l
}
\usage{ellipticity(minorAxis,majorAxis)}
\arguments{
  \item{minorAxis}{ minor ellipse axis}
  \item{majorAxis}{ major ellipse axis}
}
\value{
  scalar result
}
\source{
Wikipedia
Ellipse
\url{http://en.wikipedia.org/wiki/Ellipse#Mathematical_definitions_and_properties}
}
\references{
Wikipedia
Ellipse
\url{http://en.wikipedia.org/wiki/Ellipse#Mathematical_definitions_and_properties}
}
\author{Jose Gama}
\keyword{datasets}
